docs(metrics): snippets split, improved, and lint
