+++ examples/logger/src/append_and_remove_keys_output.json
