+++ examples/logger/src/set_correlation_id_method.json
